Move frame_to_string from types to streaming.h
