+++ src/core/streaming.h
@@ -174,6 +174,15 @@ namespace librealsense
         return os;
     }
 
+    std::string frame_holder_to_string(const frame_holder & f);
+
+    std::string frame_to_string(const frame_interface & f);
+
+    inline std::ostream& operator<<(std::ostream& out, const frame_interface & f)
+    {
+        return out << frame_to_string(f);
+    }
+
     class recommended_proccesing_blocks_interface
     {
     public:
